Time: 7 ms (100%), Space: 7.5 MB (100%) - LeetHub

+++ 2811-determine-the-minimum-sum-of-a-k-avoiding-array/2811-determine-the-minimum-sum-of-a-k-avoiding-array.cpp
@@ -4,7 +4,7 @@ public:
 
         std::vector<int> result;
         int i = 1;
-        while (result.size() < static_cast<size_t>(n)) {
+        while (result.size() < n) {
             bool shouldAdd = true;
             for (int num : result) {
                 if (i + num == k) {
